biblio_app/requirements.txt
django==3.2.*
django-rest-framework
Pillow

